Move Data Layer related classes to their own package
Separate classes specifically managing the connection between phone and Wear
