+++ wearable/src/main/java/com/google/android/apps/muzei/datalayer/ArtworkChangedListenerService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.apps.muzei;
+package com.google.android.apps.muzei.datalayer;
 
 import android.content.Intent;
